RFID Arduino door lock
In this project a servo driven lock system is made which can be locked and unlocked by using an RFID tag. An Arduino board along with an RC522 RFID sensor is used. The raid tag reads the tag and if the raid codes matches with registered code then the RFID tag is authenticated and servo motor unlocks the door lock. The servo motor automatically locks the door automatically after 5 seconds. Also Includes a 12C LCD display to indicates status of the system.